Cleanup `ReanimatedSwipeable` (#3579)
## Changes

- Replace `forwardRef` with prop `ref`
- Fix possible infinite re-render bug
- Fix invalid dependency lists
- Split `ReanimatedSwipeable` into 3 smaller files: 
  - `index.ts`
  - `ReanimatedSwipeable.tsx`
  - `ReanimatedSwipeableProps.tsx`

No changes to the core logic have been made.

## Test plan

Use available Swipeable examples to test if it works.
